feat(snippet): implement basic autocompletion for snippets --disabled
Description: this commit implements an in-process lsp in neovim to
handle snippets. Moreover, it implements a simple function to make json
snippets compatible with the lsp protocol. Lastly, some helper function
are used to read files and launch the server.

Problems: generally the autocompletion and snippet implementation is
buggy:
- If no $0 is used the completion cannot go through
- I think the score is not respected for correct ordering of results
- Sometimes the the natural server results do not pop
- The autotriggering is weird and requires some fine-tuning to get right
what keys to use

PS: This is just a hack, if vim.lsp.server() gets merged this will be
more streamlined

Signed-off-by: Dimitris Dimitropoulos
<dimitris.dimitropoulos00@gmail.com>