feat(snippet): add multiple sources for all filetypes
Problem: with the previous setup using multiple json files was a hassle.
Moreover, the json file was meant to be added manually.

Solution: add a function to parse the package.json file to get the json
files to add. Then in the configuration part amend all the files to
launch only one server. In addition, the parsing function now accepts a
filetype, enabling vim.bo.filetype utilization.