feat(snippet): implement on the fly adjustments
Problem: the server can not adapt when changing buffers.

Solution: add logic to stop and start the server based. In this commit
the keymaps for the snippet jumpstops are also moved, in order to
simplify the autocommand. Moreover, the snippet setup has also turned
more prose for debugging and understanding purposes.

Future: a signal 15 error is thrown when deleting the buffer and moving
into a new one, it has to do with the way the server exits, for now this
part in the main snippet module is commented out. Ideally, it shall
function with no error